apps/ai-orchestrator/requirements.txt
pydantic-ai
httpx
python-dotenv
fastapi
uvicorn